requirements.txt
# requirements.txt

torch
transformers
gradio
numpy
Pillow
matplotlib
